% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspGenerics.R
\name{as.data.frame.betaStacyProcess}
\alias{as.data.frame.betaStacyProcess}
\title{Converts the bsp output (support, centeringMeasure, precision) from a bsp object to a data frame}
\usage{
\method{as.data.frame}{betaStacyProcess}(x, row.names = NULL, optional = FALSE, alphaAfter = TRUE, ...)
}
\arguments{
\item{x}{The bsp object to get the output from}

\item{row.names}{Optional row names for the new data frame.}

\item{optional}{Part of as.data.frame() signature. Ignored.}

\item{alphaAfter}{Should the precision be evaluated just after time t? Default TRUE}

\item{...}{further arguments to be passed to or from methods}
}
\value{
A data frame of parameters for the given bsp object
}
\description{
Converts the bsp output (support, centeringMeasure, precision) from a bsp object to a data frame
}
\details{
This function Converts the bsp output (support, centeringMeasure, precision) from a bsp object to a data frame.
}
\note{
The precision returned by this object is the precision JUST AFTER time t.
}
