% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspSampling.R
\name{bspConfint}
\alias{bspConfint}
\title{Finds approximate (simulated) credible intervals for the percent failed at a given time}
\usage{
bspConfint(
  bsp,
  times = NULL,
  cred.level = 0.95,
  reps = 1000,
  round = -1,
  cols = TRUE
)
}
\arguments{
\item{bsp}{The bsp object}

\item{times}{Place where you would like credible intervals. If none provided, the default is the entire support of the bsp object.}

\item{cred.level}{Credible level for intervals (default is 95\%)}

\item{reps}{Number of samples at each time point used in calculations for the credible interval. Default is 1000. If the bsp object already has samples stored in bsp$Samples, reps is ignored.}

\item{round}{Number of decimal places to round the respective times to. A value of -1 means no rounding. Default is -1}

\item{cols}{(Depricated) Output credible intervals on the columns? (result will be a length(times)x3 data frame with times also as a column). 
Default is TRUE. To get credible interval results similar to package version 0.1.0, use FALSE.}
}
\value{
A length(times)x3 data frame with the support, lower, and upper bounds of the credible interval on the columns.
}
\description{
Finds approximate (simulated) credible intervals for the percent failed at a given time
}
\note{
If you intend to call this function multiple times on the same object,
changing the genSamples arguement to TRUE in the bsp() object or calling
bsp$Samples<-bspSampling(bsp) will improve performance.
}
\examples{
bsp=bsp(c(1:3), centeringMeasure = c(.1,.9, .98), precision = 2)
bspConfint(bsp, 1:3)

}
