% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspHelper.R
\name{bspPrior}
\alias{bspPrior}
\title{Creates a BSP prior from some data. User is able to shift/scale the prior and set its precision.}
\usage{
bspPrior(
  data,
  EF = 0,
  centering.max = NULL,
  prior.prec = 1,
  suppress.messages = FALSE
)
}
\arguments{
\item{data}{The data which to create the prior from. Needs to be a matrix with 2 columns for the failure times and censored indicator}

\item{EF}{A modified Effectiveness Factor (more info: http://reliawiki.org/index.php/Crow_Extended). A number between -1 and 1.
A value of 0 indicates no shifting in the support. A non-negative value (EF) indicates a system that's 1/(1-EF) as good. A negative value
indicates a system that is 1/(1-|EF|) times as worse. Mathematically, this scales the support by 1/(1-EF) if EF is positive
or 1-|EF| if EF is negative. Default is 0 (or no scaling)}

\item{centering.max}{Rescales all centering measure values such that the centering measure at the maximum support is centering.max.
Default is NULL, which means no rescaling.}

\item{prior.prec}{A single number or vector indicating the precision to set for the bsp() object. Recall that this is the precision
just after each support point. Default is 1. If NULL, use the precisions generated from a zero precision prior with the data provided.}

\item{suppress.messages}{If prior.prec is NULL, Sometimes messages are printed due to a zero precision prior. The default is FALSE, 
but if this is TRUE, it suppress them.}
}
\value{
A scaled or shifted bsp object created from provided data. Can be used as a prior.
}
\description{
Creates a BSP prior from some data. User is able to shift/scale the prior and set its precision.
}
\examples{
d<-matrix(c(5,1,15, 1, 31,1, 2,0, 20, 0), ncol=2, byrow=TRUE)
bspPrior(d, EF=2/3, prior.prec=0.5) # Three times as good

}
