% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspSampling.R
\name{bspSampling}
\alias{bspSampling}
\title{Creates a matrix where each row represents a sampled CDF from the provided BSP}
\usage{
bspSampling(bsp, reps = 10000, seed = 0)
}
\arguments{
\item{bsp}{Any bsp object}

\item{reps}{Number of samples to draw, default 10000}

\item{seed}{Seet to set before generating the samples. If NULL, don't set the seed. Deafult 0.}
}
\value{
A matrix. Each row can be plotted against
the BSP's support to view a sampled CDF from the BSP.
Column names are the support point for that column. Use samplesAt to
get extract sampled distribution at single point
}
\description{
Creates a matrix where each row represents a sampled CDF from the provided BSP
}
\examples{
bsp<-bsp(c(1:3), centeringMeasure = c(.1,.5, .98), precision = 4)
samples<-bspSampling(bsp, reps=200)
#To visualize
matplot(matrix(bsp$support[-1], nrow=3), t(samples), type='l', xlab='support', ylab='Centering Measure')
}
