% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspGenerics.R
\name{bspSdFailureTime}
\alias{bspSdFailureTime}
\title{Gives the standard deviation of time to failure of a BSP object}
\usage{
bspSdFailureTime(bsp)
}
\arguments{
\item{bsp}{The bsp object to calculate from}
}
\value{
The standard deviation of time to failure
}
\description{
Gives the standard deviation of time to failure of a BSP object
}
\details{
This function calculates the standard deviation of the failure time for a given BSP object
}
