% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsp.R
\name{evaluate_centering_measure}
\alias{evaluate_centering_measure}
\title{Evaluate centering measure at specific times}
\usage{
evaluate_centering_measure(
  bsp,
  times = NULL,
  round = 4,
  linear.interpolation = FALSE
)
}
\arguments{
\item{bsp}{A BSP object}

\item{times}{An optional vector of times that the centering measure should be determined for. If none provided, the default is over the entire support of the bsp object.}

\item{round}{Number of decimal places to round the respective times to. A value of -1 means no rounding. Default is 4}

\item{linear.interpolation}{Should linear interpolation be used to evalue the centering measure between support points? Note: this will not work past the edge of the bsp support.}
}
\value{
A named vector of same length as times with the centering measure for each time. The names correspond to the times.
}
\description{
Evaluate centering measure at specific times
}
\note{
For times in between jumps on the support, the centering measure is
considered equal to its value after the last jump
}
\examples{
evaluate_centering_measure(bsp(c(1,2), c(.2,.6), 1), c(.5,1.5,2.5))
}
