% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsp.R
\name{evaluate_precision}
\alias{evaluate_precision}
\title{Evaluate precision at specific times}
\usage{
evaluate_precision(bsp, times = NULL, round = 4, alphaAfter = FALSE)
}
\arguments{
\item{bsp}{A BSP object}

\item{times}{An optional vector of times that the precesion should be determined for. If none provided, the default is over the entire support of the bsp object.
Default is over the entire support of the bsp.}

\item{round}{Number of decimal places to round the respective times to. A value of -1 means no rounding. Default is 4}

\item{alphaAfter}{If TRUE, the returned value will reflect the precision just after the given support times. Default is
FALSE. Recall that the precision function is left continous.}
}
\value{
A named vector of same length as times with the precision for each time. The names correspond to the times.
}
\description{
Evaluate precision at specific times
}
\note{
IMPORTANT: The precision function is left continous, unlike the centering measure. For computational efficiency, the precision stored
with the bsp class (bsp$precision) gives the precision value epsilon after time t. The evaluate_precision() function gives the precision at exactly time t, unless alphaAfter=TRUE.
}
\examples{
evaluate_precision(bsp(c(1,2), c(.2,.6), 1), c(.5,1.5,2.5))
}
