% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspGenerics.R
\name{mean.betaStacyProcess}
\alias{mean.betaStacyProcess}
\title{Gives the mean time to failure of a BSP object}
\usage{
\method{mean}{betaStacyProcess}(x, ...)
}
\arguments{
\item{x}{The bsp object to calculate from}

\item{...}{further arguments to be passed to or from methods}
}
\value{
Mean time to Failure
}
\description{
Gives the mean time to failure of a BSP object
}
\details{
This function calculates the expected failure time for a given BSP object
}
