% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspGenerics.R
\name{median.betaStacyProcess}
\alias{median.betaStacyProcess}
\title{Gives the median time to failure of a BSP object}
\usage{
\method{median}{betaStacyProcess}(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{The bsp object to calculate from}

\item{na.rm}{Part of median() signature. Ignored.}

\item{...}{further arguments to be passed to or from methods}
}
\value{
Median time to Failure
}
\description{
Gives the median time to failure of a BSP object
}
\details{
This function calculates the median failure time for a given BSP object
}
