% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspGenerics.R
\name{plot.betaStacyProcess}
\alias{plot.betaStacyProcess}
\title{Plots a Beta-Stacy process over its support}
\usage{
\method{plot}{betaStacyProcess}(
  x,
  withConfInt = FALSE,
  ncdfs = FALSE,
  cred.level = NULL,
  withLegend = T,
  reps = 1000,
  title = "",
  xlim = NULL,
  xlab = "Time (t)",
  ylab = "Centering Measure G(t)",
  ...
)
}
\arguments{
\item{x}{The bsp object to be plotted}

\item{withConfInt}{If true, will calculate and plot approximate 95\% confidence intervals, default FALSE}

\item{ncdfs}{Plots ncdfs CDFs in gray. If FALSE or 0, no cdfs will be plotted. Default is FALSE}

\item{cred.level}{Level for credible interval, if none provided, default is 0.95}

\item{withLegend}{Only when withConfInt=TRUE. Should the legend be included on the plot?}

\item{reps}{Only valid when withConfint is TRUE. Generates max(reps, ncdfs) samples. Default is 1000}

\item{title}{Title for the plot. Default is an empty string}

\item{xlim}{X-limits of the plot}

\item{xlab}{label for the x axis. Default is "Time (t)"}

\item{ylab}{label for the y axis. Defualt is "Centering Measure G(t)"}

\item{...}{further arguments to be passed to or from methods}
}
\value{
A ggplot object
}
\description{
Plots a Beta-Stacy process over its support
}
\examples{
plot(bsp(c(1,5,10),c(0.2,0.4, .8),1))


}
