% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspHelper.R
\name{plotBspGrid}
\alias{plotBspGrid}
\title{Plots a list of bsp objects in a grid-like pattern}
\usage{
plotBspGrid(
  l,
  nrow = NULL,
  ncol = NULL,
  titles = paste("Comp", 1:length(l)),
  xlims = rep(NULL, length(l)),
  priors = NULL,
  data = NULL,
  prior.col = "blue",
  ...
)
}
\arguments{
\item{l}{The list of bsp objects to be plotted. These are plotted row-wise.}

\item{nrow}{(Optional) Number of rows in the plotted grid}

\item{ncol}{(Optional) Number of columns in the plotted grid}

\item{titles}{A vector of titles for the plots. Default paste('Comp', 1:length(l))}

\item{xlims}{list of x-axis limits to use for each respective plot. If an element is NULL, make no changes to the xlimits. 
The default is a list of NULLs.}

\item{priors}{Optional list of priors that are plotted on top of the bsp objects. If provided, it will plot a blue line indicating the prior centering measure. Note:
This only works as long as each prior is explicitly defined (including any subsystem components).}

\item{data}{Optional list of data matrices who's kaplan-meier curves are plotted on top of the bsp objects. If provided, it will plot dashed black lines for each object.}

\item{prior.col}{Optional color string for the prior line. Only used if priors is provided.}

\item{...}{further arguments to be passed to the plot.betaStacyProcess() function. Allows the user to suppress the legend or add confidence bands, for example.}
}
\value{
A gtable grid of bsp ggplots.
}
\description{
Plots a list of bsp objects in a grid-like pattern
}
\note{
This function requires the gridExtra package. If the data is plotted, it requires the Survival package.
}
\examples{
set.seed(0)
C.A <- cbind(rweibull(15, 3, 10), rbinom(15, 1, .6))
C.B <- cbind(rweibull(10, 4, 10), rbinom(10, 1, .6))
C.C <- cbind(rweibull(20, 5, 10), rbinom(20, 1, .6))
C <- list(C.A, C.B, C.C)

ps <- lapply(1:length(C), function(x) bsp(c(8,9,10,11), c(.1, .3, .6, .9), 5))
sys.names <- paste0(paste0('c',1:3))
names(ps) <- names(C) <- sys.names
txt <- paste0('S','(',paste0('c',1:3, collapse = ',') ,'):System')
post <- estimateSystemReliability(textConnection(txt), priorList = ps, dataList = C)

plotBspGrid(post[1:3], ncol=2, withConfInt=TRUE, withLegend=FALSE, priors=ps, data=C, ylab='G(t)')

}
