% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspGenerics.R
\name{print.betaStacyProcess}
\alias{print.betaStacyProcess}
\title{Prints a Beta-Stacy process object as a data frame}
\usage{
\method{print}{betaStacyProcess}(x, alphaAfter = TRUE, ...)
}
\arguments{
\item{x}{The bsp object to be printed}

\item{alphaAfter}{Should the precision be evaluated just after time t? Default TRUE}

\item{...}{further arguments to be passed to or from methods}
}
\value{
none
}
\description{
Prints a Beta-Stacy process object as a data frame
}
\note{
The precision printed by this object is the precision JUST AFTER time t, unless alphaAfter is FALSE.
}
\examples{
print(bsp(c(1,5),c(0.2,0.4),1))


}
