% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspGenerics.R
\name{summary.betaStacyProcess}
\alias{summary.betaStacyProcess}
\title{Gives summary statistics of time to failure of a BSP object}
\usage{
\method{summary}{betaStacyProcess}(object, quants = c(0.05, 0.1, 0.25, 0.5, 0.75, 0.9, 0.95), ...)
}
\arguments{
\item{object}{The bsp object to calculate from}

\item{quants}{Optional failure time quantiles to include in the summary. 
Default is c(.05, .1, .25, .5, .75, .9, .95)}

\item{...}{further arguments to be passed to or from methods}
}
\value{
Summary statistics for the time to failure
}
\description{
Gives summary statistics of time to failure of a BSP object
}
\details{
This function calculates summary statistics of the failure time for a given BSP object
}
